+++ CHANGES.md
@@ -81,6 +81,7 @@ Changes in 9.0
 - Restrict diagnostic access to logged in admin in diagnostic.php, thanks to @intrapus
 - Fix SQL error value too long for type character varying(50) in Schools.php
 - Add Secure RosarioSIS link in INSTALL.md
+- Add Calendar days legend in Calendar.php
 
 Changes in 8.9.6
 ----------------